<update> inventory player

+++ game-server/src/main/java/nro/server/model/ecs/component/item/ItemInfoComponent.java
@@ -7,7 +7,7 @@ import com.artemis.Component;
  */
 public class ItemInfoComponent extends Component {
 
-    public int templateId;
+    public short templateId;
     public int quantity;
     public int creator_id = -1;
 
@@ -15,7 +15,7 @@ public class ItemInfoComponent extends Component {
     }
 
     public ItemInfoComponent(int templateId, int quantity, int creator_id) {
-        this.templateId = templateId;
+        this.templateId = (short) templateId;
         this.quantity = quantity;
         this.creator_id = creator_id;
     }
