+++ game-server/src/main/java/nro/server/network/nro/server_packets/handler/SmMapInfo.java
@@ -53,9 +53,9 @@ public class SmMapInfo extends NroServerPacket {
         writeByte(0);
 
         // send background map
-        writeByte(0);
+        writeShort(0);
         // send effect map
-        writeByte(0);
+        writeShort(0);
 
         writeByte(mapTemplate.getBgType());
         writeByte(positionComponent.teleport);
