+++ game-server/src/main/java/nro/server/world/WorldPosition.java
@@ -3,5 +3,5 @@ package nro.server.world;
 /**
  * @author Arriety
  */
-public record WorldPosition(short mapId, int x, int y, int zoneId, WorldMapInstance parent) {
+public record WorldPosition(short mapId, short x, short y, int zoneId, WorldMapInstance parent) {
 }
